core/images/request-controller/src/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
pydantic==2.10.6
pydantic_core==2.27.2
kafka==1.3.5
httpx==0.28.1
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.29.3
prometheus_fastapi_instrumentator==7.0.2
requests==2.31.0